Move `cuda::aligned_size` to its own file
